+++ stralg/aho_corasick.h
@@ -3,29 +3,30 @@
 
 #include <stddef.h>
 #include <trie.h>
+#include <stdint.h>
 
 struct ac_iter {
     const char *text;
-    size_t n;
+    uint32_t n;
 
-    const size_t *pattern_lengths;
+    const uint32_t *pattern_lengths;
     struct trie *patterns_trie;
 
     bool nested;
-    size_t j;
+    uint32_t j;
     struct trie *v, *w;
     struct output_list *hits;
 };
 
 struct ac_match {
     int string_label;
-    size_t index;
+    uint32_t index;
 };
 void init_ac_iter(
     struct ac_iter *iter,
     const char *text,
-    size_t n,
-    const size_t *pattern_lengths,
+    uint32_t n,
+    const uint32_t *pattern_lengths,
     struct trie *patterns_trie
 );
 bool next_ac_match(
