+++ stralg/match.c
@@ -9,8 +9,8 @@
 
 void init_naive_match_iter(
     struct naive_match_iter *iter,
-    const char *text, size_t n,
-    const char *pattern, size_t m
+    const char *text, uint32_t n,
+    const char *pattern, uint32_t m
 ) {
     assert(m > 0);
     
@@ -27,12 +27,12 @@ bool next_naive_match(
     struct naive_match_iter *iter,
     struct match *match
 ) {
-    size_t n = iter->n, m = iter->m;
+    uint32_t n = iter->n, m = iter->m;
     const char *text = iter->text;
     const char *pattern = iter->pattern;
 
-    for (size_t j = iter->current_index; j <= n - m; j++) {
-        size_t i = 0;
+    for (uint32_t j = iter->current_index; j <= n - m; j++) {
+        uint32_t i = 0;
         while (i < m && text[j+i] == pattern[i]) {
             i++;
         }
@@ -55,8 +55,8 @@ void dealloc_naive_match_iter(
 
 void init_border_match_iter(
     struct border_match_iter *iter,
-    const char *text, size_t n,
-    const char *pattern, size_t m
+    const char *text, uint32_t n,
+    const char *pattern, uint32_t m
 ) {
     assert(m > 0);
     
@@ -64,7 +64,7 @@ void init_border_match_iter(
     iter->pattern = pattern; iter->m = m;
     iter->i = iter->b = 0;
 
-    size_t *ba = malloc(m * sizeof(size_t));
+    uint32_t *ba = malloc(m * sizeof(uint32_t));
     ba[0] = 0;
     for (int i = 1; i < m; ++i) {
         int b = ba[i - 1];
@@ -81,11 +81,11 @@ bool next_border_match(
 ) {
     const char *text = iter->text;
     const char *pattern = iter->pattern;
-    size_t *ba = iter->border_array;
-    size_t b = iter->b;
-    size_t m = iter->m;
+    uint32_t *ba = iter->border_array;
+    uint32_t b = iter->b;
+    uint32_t m = iter->m;
 
-    for (size_t i = iter->i; i < iter->n; ++i) {
+    for (uint32_t i = iter->i; i < iter->n; ++i) {
         while (b > 0 && text[i] != pattern[b])
             b = ba[b - 1];
         b = (text[i] == pattern[b]) ? b + 1 : 0;
@@ -135,8 +135,8 @@ static void ba_search(char * key, char * buffer)
 
 void init_kmp_match_iter(
     struct kmp_match_iter *iter,
-    const char *text, size_t n,
-    const char *pattern, size_t m
+    const char *text, uint32_t n,
+    const char *pattern, uint32_t m
 ) {
     assert(m > 0);
     
@@ -152,10 +152,10 @@ void init_kmp_match_iter(
     // Build prefix border array -- I allocate with calloc
     // because the static analyser otherwise think it can contain
     // garbage values after the initialisation.
-    size_t *prefixtab = calloc(m, sizeof(size_t));
+    uint32_t *prefixtab = calloc(m, sizeof(uint32_t));
     prefixtab[0] = 0;
-    for (size_t i = 1; i < m; ++i) {
-        size_t k = prefixtab[i - 1];
+    for (uint32_t i = 1; i < m; ++i) {
+        uint32_t k = prefixtab[i - 1];
         while (k > 0 && pattern[i] != pattern[k])
             k = prefixtab[k - 1];
         prefixtab[i] = (pattern[i] == pattern[k]) ? k + 1 : 0;
@@ -163,7 +163,7 @@ void init_kmp_match_iter(
 
     // Modify it so the we avoid borders where the following
     // letters match
-    for (size_t i = 0; i < m - 1; i++) {
+    for (uint32_t i = 0; i < m - 1; i++) {
         prefixtab[i] =
             (pattern[prefixtab[i]] != pattern[i + 1] || prefixtab[i] == 0) ?
             prefixtab[i] : prefixtab[prefixtab[i] - 1];
@@ -177,10 +177,10 @@ bool next_kmp_match(
     struct match *match
 ) {
     // aliases to make the code easier to read...
-    size_t j = iter->j;
-    size_t i = iter->i;
-    size_t m = iter->m;
-    size_t max_match_index = iter->max_match_len;
+    uint32_t j = iter->j;
+    uint32_t i = iter->i;
+    uint32_t m = iter->m;
+    uint32_t max_match_index = iter->max_match_len;
     const char *text = iter->text;
     const char *pattern = iter->pattern;
 
@@ -220,8 +220,8 @@ void dealloc_kmp_match_iter(
 
 void init_bmh_match_iter(
     struct bmh_match_iter *iter,
-    const char *text, size_t n,
-    const char *pattern, size_t m
+    const char *text, uint32_t n,
+    const char *pattern, uint32_t m
 ) {
     assert(m > 0);
     assert(m <= n);
@@ -229,10 +229,10 @@ void init_bmh_match_iter(
     iter->j = 0;
     iter->text = text; iter->n = n;
     iter->pattern = pattern; iter->m = m;
-    for (size_t k = 0; k < 256; k++) {
+    for (uint32_t k = 0; k < 256; k++) {
         iter->jump_table[k] = m;
     }
-    for (size_t k = 0; k < m - 1; k++) {
+    for (uint32_t k = 0; k < m - 1; k++) {
         iter->jump_table[(unsigned char)pattern[k]] = m - k - 1;
     }
 }
@@ -244,15 +244,15 @@ bool next_bmh_match(
     // aliasing to make the code easier to read...
     const char *text = iter->text;
     const char *pattern = iter->pattern;
-    size_t n = iter->n;
-    size_t m = iter->m;
-    size_t *jump_table = iter->jump_table;
+    uint32_t n = iter->n;
+    uint32_t m = iter->m;
+    uint32_t *jump_table = iter->jump_table;
 
-    for (size_t j = iter->j;
+    for (uint32_t j = iter->j;
          j < n - m + 1;
          j += jump_table[(unsigned char)text[j + m - 1]]) {
 
-        size_t i = m - 1;
+        uint32_t i = m - 1;
         while (i > 0 && pattern[i] == text[j + i]) {
             i--;
         }
